python-ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.27.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.0
aiokafka==0.10.0
redis==5.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx==0.26.0
numpy==1.26.0
scikit-learn==1.3.0
python-json-logger==2.0.7
prometheus-client==0.19.0
websockets==12.0
langchain-google-genai>=2.0.0
